requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
python-dotenv>=1.0.0
pypdf>=3.9.0
sentence-transformers>=2.3.1
openai>=1.12.0
python-magic>=0.4.27
tiktoken>=0.6.0
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pinecone-client>=3.0.0
langchain-pinecone>=0.0.1 